// ng-telford-leisure-services/src/app/core/models/member.ts
export interface Member {
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  gender: string;
  email: string;
  phone: string;
  addressLineOne: string;
  addressLineTwo: string;
  townOrCity: string;
  county: string;
  postcode: string;
  ethnicity: string;
  mainCenter: string;
  membershipType: string;
}